Search passage in read mode #367 refactor code
